order-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
python-dotenv==1.0.1
strawberry-graphql[fastapi]==0.219.0
python-multipart==0.0.6
httpx==0.26.0
cryptography==41.0.7